Implemented a dropdown list for switching to and closing windows of a given application
WindowListView implements the list itself by using Gtk.Popover as the base
widget and a Gtk.ListBox to store the rows associated with the application
windows. WindowListItem implements the row. It has two buttons, one for
switching to the new window and another for closing it. These events are passed
through the WindowManager interface.